requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
tensorflow>=2.13.0
Pillow>=10.0.0
numpy>=1.24.3
